rename tensorflow_compat tensorflow_api to pose_estimation
